requirements.txt
aiosmtplib==4.0.2
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.0.1
beanie==1.26.0
black==25.9.0
cachetools==6.2.1
certifi==2024.8.30
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.2
dnspython==2.8.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.119.0
filetype==1.2.0
anthropic>=0.39.0
google-ai-generativelanguage==0.9.0
google-api-core==2.28.1
google-auth==2.41.1
google-genai==1.41.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.0.7
langchain-core==1.0.5
langchain-google-genai==3.0.3
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.4
langgraph-sdk==0.2.9
langsmith==0.4.43
lazy-model==0.2.0
motor==3.3.2
mypy_extensions==1.1.0
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
pinecone-client==6.0.0
pinecone-plugin-interface==0.0.7
platformdirs==4.5.0
proto-plus==1.26.1
protobuf==6.33.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.10.3
pydantic-settings==2.5.2
pydantic_core==2.27.1
pymongo==4.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.3.0
pytokens==0.1.10
authlib==1.3.0
google-auth==2.41.1
itsdangerous==2.2.0
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
tenacity==9.1.2
toml==0.10.2
typing-inspection==0.4.2
typing_extensions==4.15.0
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.5.0
user-agents==2.2.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
xxhash==3.6.0
zstandard==0.25.0
